text_pre_djangorest_api/text_preprocessing/api/serializers.py
from rest_framework import serializers
from .models import RegexEntry


class RegexSerializer(serializers.ModelSerializer):
    class Meta:
        model = RegexEntry
        fields = [
            'name',
            'pattern',
            'replace_token'
        ]
