text_pre_djangorest_api/text_preprocessing/requirements.txt
django
djangorestframework
requests